Downloads/Websocket-PRJCT/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
websockets==11.0.3
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4